requirements.txt
streamlit
bcrypt
pillow
reportlab
pandas
streamlit-js-eval
pytz
streamlit-aggrid
